+++ lib/uniword.rb
@@ -135,6 +135,7 @@ module Uniword
   autoload :GenerateCLI, "uniword/cli/generate_cli"
   autoload :FontsCLI, "uniword/cli/fonts_cli"
   autoload :PageCLI, "uniword/cli/page_cli"
+  autoload :StylesCLI, "uniword/cli/styles_cli"
   autoload :DiffCLI, "uniword/cli/diff_cli"
   autoload :TemplateCLI, "uniword/cli/template_cli"
   autoload :CLIHelpers, "uniword/cli/helpers"